Change arch name to be hsail64 instead of hsail-64
This needs to match the triple arch name. This was failing to match,
so this would default to the host architecture and triple. This just
happens to work correctly on Linux. For some reason, on OS X this would
result in some kind of inconsistency trying to use ELF sometimes while
set up for MachO, so hsail64 would always crash.