Remove MEMOP32, and try to avoid iPTR
iPTR is fundamentally broken. It pretends to have a pointer type
sometimes, but in general SelectionDAG loses that information.
There's no reasonable way for it to accept an address space, so just
accept any integer type for pointer operands. The custom operand's
base type is still considered to be iPTR, but I'm not sure that really
means anything with suboperands.