Fix assert by cherry picking r229443
Since we are currently using unknown for the output operand type,
this fails to pick a register class for the mov instruction. We
should probably figure out something to avoid this, since it seems
like this will break a lot of things. There maybe should be a register
class that includes both s and d registers.

Original commit message:

RegisterCoalescer: Do not look for regclass of IMPLICIT_DEF.

IMPLICIT_DEF is a generic instruction and has no (fixed) output register
class defined. The rematerialization code of the register coalescer
should not scan the instruction description for a register class.

This fixes a problem showing up in
test/CodeGen/R600/subreg-coalescer-crash.ll with subregister liveness
enabled.