HSAIL: Rename srcTypesrcLength modifier operand name
This deviates from the spec, but makes it more consistent for us to use.
InstSourceType call this sourceType in the BRIG, but the descriptions
of the instructions call it srcTypesrcLength. It's more useful for
us to just have a single name to refer to this for all instructions
instead of trying to use both in different places.'