HSAIL: Fix broken behavior handling i1 arguments
This was deciding that sometimes i1 values needed to be treated
as i32, which is inconsistent with what everything else expected.
This causes various problems, such as an assert when lowering parameters
with -debug. If any frontends were relying on this for ABI reasons,
they will need to adjust the emitted argument types.