HSAIL: Fix an erroneus negation of a condition inside a loop.
Now the given test case passes, but likely a more complex case with
the use in a farther BB in the loop would not.

An alternative fix would be to ensure the _negation instruction_
will be inserted outside the loop (in case the original condition is
not needed), immediately after the producer of the condition.
Now the not was inserted inside a loop header which caused loop
control flow breakage (branch taken every second time).